move UpdateCWDtoManifestPath to filesystem
