+++ cmd/deploy/deploy.go
@@ -840,7 +840,7 @@ func checkOktetoManifestPathFlag(options *Options, fs afero.Fs) error {
 		options.ManifestPathFlag = manifestPathFlag
 
 		// when the manifest path is set by the cmd flag, we are moving cwd so the cmd is executed from that dir
-		uptManifestPath, err := model.UpdateCWDtoManifestPath(options.ManifestPath)
+		uptManifestPath, err := filesystem.UpdateCWDtoManifestPath(options.ManifestPath)
 		if err != nil {
 			return err
 		}
