+++ cmd/destroy/destroy.go
@@ -17,6 +17,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/okteto/okteto/pkg/filesystem"
 	"os"
 	"path/filepath"
 	"strings"
@@ -154,7 +155,7 @@ func Destroy(ctx context.Context, at analyticsTrackerInterface, insights buildTr
 				options.ManifestPathFlag = manifestPathFlag
 
 				// when the manifest path is set by the cmd flag, we are moving cwd so the cmd is executed from that dir
-				uptManifestPath, err := model.UpdateCWDtoManifestPath(options.ManifestPath)
+				uptManifestPath, err := filesystem.UpdateCWDtoManifestPath(options.ManifestPath)
 				if err != nil {
 					return err
 				}
