+++ pkg/filesystem/workdir_test.go
@@ -14,6 +14,7 @@
 package filesystem
 
 import (
+	"os"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -123,6 +124,69 @@ func Test_GetManifestPathFromWorkdir(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			res := GetManifestPathFromWorkdir(tt.path, tt.workdir)
 			assert.Equal(t, tt.expectedPath, res)
+		})
+	}
+}
+
+func Test_UpdateCWDtoManifestPath(t *testing.T) {
+	root := t.TempDir()
+	var tests = []struct {
+		name         string
+		path         string
+		expectedPath string
+		expectedErr  bool
+	}{
+		{
+			name:         "inside .okteto folder",
+			path:         filepath.Join(root, ".okteto", "okteto.yml"),
+			expectedPath: filepath.Join(".okteto", "okteto.yml"),
+		},
+		{
+			name:         "one path ahead",
+			path:         filepath.Join(root, "test", "okteto.yml"),
+			expectedPath: "okteto.yml",
+		},
+		{
+			name:         "same path",
+			path:         filepath.Join(root, "okteto.yml"),
+			expectedPath: "okteto.yml",
+		},
+		{
+			name:         "full path on .okteto",
+			path:         filepath.Join(root, "usr", ".okteto", "okteto.yml"),
+			expectedPath: filepath.Join(".okteto", "okteto.yml"),
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+
+			initialCWD, err := os.Getwd()
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if err := os.Chdir(root); err != nil {
+				t.Fatal(err)
+			}
+
+			if err := os.MkdirAll(tt.path, 0700); err != nil {
+				t.Fatal(err)
+			}
+
+			res, err := UpdateCWDtoManifestPath(tt.path)
+			if tt.expectedErr && err == nil {
+				t.Fatal("expected err")
+			}
+			if !tt.expectedErr && err != nil {
+				t.Fatalf("not expected error, got %v", err)
+			}
+
+			assert.Equal(t, tt.expectedPath, res)
+
+			if err := os.Chdir(initialCWD); err != nil {
+				t.Fatal(err)
+			}
 		})
 	}
 }
